refactor: colocate admin quest management tests
