+++ components/admin/quest-management-tab.sections.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import { QuestStatus } from "@/lib/types/database";
-import { QuestManagementTab } from "../quest-management-tab";
+import { QuestManagementTab } from "./quest-management-tab";
 
 // jest.mock calls must be in the test file itself to be hoisted properly
 jest.mock("@/hooks/useQuests", () => ({
